--- a/table.go
+++ b/table.go
@@ -720,13 +720,50 @@ func (t *TableBlock) Insert(ctx context.Context, tx uint64, buf *dynparquet.Seri
 		return fmt.Errorf("failed to split rows by granule: %w", err)
 	}
 
+	b := bytes.NewBuffer(nil)
+	w, err := t.table.config.schema.GetWriter(b, buf.DynamicColumns())
+	if err != nil {
+		return fmt.Errorf("failed to get writer: %w", err)
+	}
+
 	parts := []*Part{}
-	for granule, serBuf := range rowsToInsertPerGranule {
+	for granule, indicies := range rowsToInsertPerGranule {
 		select {
 		case <-ctx.Done():
 			tombstone(parts)
 			return ctx.Err()
 		default:
+
+			rowBuf := make([]parquet.Row, 1)
+			rows := buf.Reader()
+			for index := 0; ; index++ {
+				_, err := rows.ReadRows(rowBuf)
+				if err == io.EOF {
+					break
+				}
+				if err != nil {
+					return err
+				}
+
+				// Check if this index belongs in this granule
+				if _, ok := indicies[index]; !ok {
+					continue
+				}
+
+				_, err = w.WriteRows(rowBuf)
+				if err != nil {
+					return fmt.Errorf("failed to write rows: %w", err)
+				}
+			}
+			if err := w.Close(); err != nil {
+				return fmt.Errorf("failed to close writer: %w", err)
+			}
+
+			serBuf, err := dynparquet.ReaderFromBytes(b.Bytes())
+			if err != nil {
+				return fmt.Errorf("failed to get reader from bytes: %w", err)
+			}
+
 			part := NewPart(tx, serBuf)
 			size, err := granule.AddPart(part)
 			if err != nil {
@@ -738,6 +775,9 @@ func (t *TableBlock) Insert(ctx context.Context, tx uint64, buf *dynparquet.Seri
 				go t.compact(granule)
 			}
 			t.size.Add(serBuf.ParquetFile().Size())
+
+			b = bytes.NewBuffer(nil)
+			w.Reset(b)
 		}
 	}
 
@@ -971,57 +1011,20 @@ func (t *TableBlock) Index() *btree.BTree {
 	return (*btree.BTree)(t.index.Load())
 }
 
-func (t *TableBlock) splitRowsByGranule(buf *dynparquet.SerializedBuffer) (map[*Granule]*dynparquet.SerializedBuffer, error) {
-	// Special case: if there is only one granule, insert parts into it until full.
+func (t *TableBlock) splitRowsByGranule(buf *dynparquet.SerializedBuffer) (map[*Granule]map[int]struct{}, error) {
 	index := t.Index()
 	if index.Len() == 1 {
-		b := bytes.NewBuffer(nil)
-
-		cols := buf.DynamicColumns()
-		w, err := t.table.config.schema.GetWriter(b, cols)
-		if err != nil {
-			return nil, ErrCreateSchemaWriter{err}
-		}
-		defer t.table.config.schema.PutWriter(w)
-
-		rowBuf := make([]parquet.Row, 1)
-		rows := buf.Reader()
-		for {
-			_, err := rows.ReadRows(rowBuf)
-			if err == io.EOF {
-				break
-			}
-			if err != nil {
-				return nil, ErrReadRow{err}
-			}
-			_, err = w.WriteRows(rowBuf)
-			if err != nil {
-				return nil, ErrWriteRow{err}
-			}
+		rows := map[int]struct{}{}
+		for i := 0; i < int(buf.NumRows()); i++ {
+			rows[i] = struct{}{}
 		}
 
-		err = w.Close()
-		if err != nil {
-			return nil, fmt.Errorf("failed to close schema writer: %w", err)
-		}
-
-		serBuf, err := dynparquet.ReaderFromBytes(b.Bytes())
-		if err != nil {
-			return nil, fmt.Errorf("failed to create dynparquet reader: %w", err)
-		}
-
-		return map[*Granule]*dynparquet.SerializedBuffer{
-			index.Min().(*Granule): serBuf,
+		return map[*Granule]map[int]struct{}{
+			index.Min().(*Granule): rows,
 		}, nil
 	}
 
-	writerByGranule := map[*Granule]*dynparquet.PooledWriter{}
-	bufByGranule := map[*Granule]*bytes.Buffer{}
-	defer func() {
-		for _, w := range writerByGranule {
-			t.table.config.schema.PutWriter(w)
-		}
-	}()
+	rowsByGranule := map[*Granule]map[int]struct{}{}
 
 	// TODO: we might be able to do ascend less than or ascend greater than here?
 	rows := buf.DynamicRows()
@@ -1033,6 +1036,7 @@ func (t *TableBlock) splitRowsByGranule(buf *dynparquet.SerializedBuffer) (map[*
 	if err != nil {
 		return nil, ErrReadRow{err}
 	}
+	idx := 0
 	row := rowBuf.Get(0)
 
 	var ascendErr error
@@ -1045,22 +1049,14 @@ func (t *TableBlock) splitRowsByGranule(buf *dynparquet.SerializedBuffer) (map[*
 			isLess := t.table.config.schema.RowLessThan(row, least)
 			if isLess {
 				if prev != nil {
-					w, ok := writerByGranule[prev]
+					_, ok := rowsByGranule[prev]
 					if !ok {
-						b := bytes.NewBuffer(nil)
-						w, err = t.table.config.schema.GetWriter(b, buf.DynamicColumns())
-						if err != nil {
-							ascendErr = ErrCreateSchemaWriter{err}
-							return false
-						}
-						writerByGranule[prev] = w
-						bufByGranule[prev] = b
+						// TODO
+						rowsByGranule[prev] = map[int]struct{}{idx: struct{}{}}
+					} else {
+						rowsByGranule[prev][idx] = struct{}{}
 					}
-					_, err = w.WriteRows([]parquet.Row{row.Row})
-					if err != nil {
-						ascendErr = ErrWriteRow{err}
-						return false
-					}
+
 					n, err := rows.ReadRows(rowBuf)
 					if err == io.EOF && n == 0 {
 						// All rows accounted for
@@ -1072,6 +1068,7 @@ func (t *TableBlock) splitRowsByGranule(buf *dynparquet.SerializedBuffer) (map[*
 						return false
 					}
 					row = rowBuf.Get(0)
+					idx++
 					continue
 				}
 			}
@@ -1087,23 +1084,14 @@ func (t *TableBlock) splitRowsByGranule(buf *dynparquet.SerializedBuffer) (map[*
 	}
 
 	if !exhaustedAllRows {
-		w, ok := writerByGranule[prev]
+		_, ok := rowsByGranule[prev]
 		if !ok {
-			b := bytes.NewBuffer(nil)
-			w, err = t.table.config.schema.GetWriter(b, buf.DynamicColumns())
-			if err != nil {
-				return nil, ErrCreateSchemaWriter{err}
-			}
-			writerByGranule[prev] = w
-			bufByGranule[prev] = b
+			rowsByGranule[prev] = map[int]struct{}{}
 		}
 
 		// Save any remaining rows that belong into prev
 		for {
-			_, err = w.WriteRows([]parquet.Row{row.Row})
-			if err != nil {
-				return nil, ErrWriteRow{err}
-			}
+			rowsByGranule[prev][idx] = struct{}{}
 			n, err := rows.ReadRows(rowBuf)
 			if err == io.EOF && n == 0 {
 				break
@@ -1111,22 +1099,12 @@ func (t *TableBlock) splitRowsByGranule(buf *dynparquet.SerializedBuffer) (map[*
 			if err != nil && err != io.EOF {
 				return nil, ErrReadRow{err}
 			}
+			row = rowBuf.Get(0)
+			idx++
 		}
 	}
 
-	res := map[*Granule]*dynparquet.SerializedBuffer{}
-	for g, w := range writerByGranule {
-		err := w.Close()
-		if err != nil {
-			return nil, err
-		}
-		res[g], err = dynparquet.ReaderFromBytes(bufByGranule[g].Bytes())
-		if err != nil {
-			return nil, fmt.Errorf("failed to read from granule buffer: %w", err)
-		}
-	}
-
-	return res, nil
+	return rowsByGranule, nil
 }
 
 // compact will compact a Granule; should be performed as a background go routine.