--- a/table.go
+++ b/table.go
@@ -886,7 +886,6 @@ func (t *TableBlock) compactGranule(granule *Granule) error {
 		if err != nil {
 			return fmt.Errorf("splitting granule: %w", err)
 		}
-
 	}
 
 	// add remaining parts onto new granules
@@ -1051,8 +1050,7 @@ func (t *TableBlock) splitRowsByGranule(buf *dynparquet.SerializedBuffer) (map[*
 				if prev != nil {
 					_, ok := rowsByGranule[prev]
 					if !ok {
-						// TODO
-						rowsByGranule[prev] = map[int]struct{}{idx: struct{}{}}
+						rowsByGranule[prev] = map[int]struct{}{idx: {}}
 					} else {
 						rowsByGranule[prev][idx] = struct{}{}
 					}