![image](https://user-images.githubusercontent.com/8681572/188913971-98caaa0b-ec8d-463f-a456-45f6b66ae12d.png)

This change is in lieu of https://github.com/polarsignals/frostdb/pull/179
While that change had a large amount of memory savings, it came at the cost of significantly decreased query performance due to having to open the parquet files every time. 

This change does two things:
1. Changes compaction:
granules are compacted based on their overall size in bytes from the parquet files they hold instead of raw number of rows. If however after a granule is compacted, its new size is less than that of the setting, it does not split the granule into two but instead replaces the previous granule with the new compacted one.
2. Uses a single writer during inserts instead of a writer per granule. 

```thor@thors-MacBook-Pro frostdb % benchstat before.txt 1mbgranule.txt
name                                          old time/op    new time/op    delta
QueryTypes-10                                   13.9ms ± 0%    18.4ms ± 0%   ~     (p=1.000 n=1+1)
QueryMerge-10                                    208ms ± 0%      19ms ± 0%   ~     (p=1.000 n=1+1)
QueryRange-10                                    206ms ± 0%      13ms ± 0%   ~     (p=1.000 n=1+1)
_Table_Insert_1000Rows_10Iters_10Writers-10      563ms ± 0%     597ms ± 0%   ~     (p=1.000 n=1+1)
_Table_Insert_100Rows_1000Iters_1Writers-10      539ms ± 0%     160ms ± 0%   ~     (p=1.000 n=1+1)
_Table_Insert_100Rows_100Iters_100Writers-10     68.6s ± 0%     70.7s ± 0%   ~     (p=1.000 n=1+1)

name                                          old alloc/op   new alloc/op   delta
QueryTypes-10                                   37.8MB ± 0%    26.3MB ± 0%   ~     (p=1.000 n=1+1)
QueryMerge-10                                    275MB ± 0%      54MB ± 0%   ~     (p=1.000 n=1+1)
QueryRange-10                                    260MB ± 0%      25MB ± 0%   ~     (p=1.000 n=1+1)
_Table_Insert_1000Rows_10Iters_10Writers-10     97.3MB ± 0%    55.3MB ± 0%   ~     (p=1.000 n=1+1)
_Table_Insert_100Rows_1000Iters_1Writers-10      742MB ± 0%     208MB ± 0%   ~     (p=1.000 n=1+1)
_Table_Insert_100Rows_100Iters_100Writers-10    2.90GB ± 0%    2.24GB ± 0%   ~     (p=1.000 n=1+1)

name                                          old allocs/op  new allocs/op  delta
QueryTypes-10                                     169k ± 0%      320k ± 0%   ~     (p=1.000 n=1+1)
QueryMerge-10                                    3.44M ± 0%     0.13M ± 0%   ~     (p=1.000 n=1+1)
QueryRange-10                                    3.38M ± 0%     0.07M ± 0%   ~     (p=1.000 n=1+1)
_Table_Insert_1000Rows_10Iters_10Writers-10       528k ± 0%      127k ± 0%   ~     (p=1.000 n=1+1)
_Table_Insert_100Rows_1000Iters_1Writers-10      11.6M ± 0%      1.2M ± 0%   ~     (p=1.000 n=1+1)
_Table_Insert_100Rows_100Iters_100Writers-10     20.8M ± 0%     12.2M ± 0%   ~     (p=1.000 n=1+1)```