@asubiotto 2022-09-13
Just a note: You can get benchstat to give you diff deltas by running the benchmarks with `-count`. It's a little easier to parse the performance improvements.